refactor(mokumo-shop): relocate platform startup helpers from services/api [#512]
Lifts ensure_data_dirs, prepare_database, init_session_and_setup,
resolve_demo_install_ok, try_bind, migrate_flat_layout,
resolve_active_profile, generate_setup_token, and ProfileDbError from
services/api/src/lib.rs into crates/mokumo-shop/src/startup.rs.

These helpers call into mokumo_shop::db functions (schema compatibility
check, pool initialization), so they cannot live in kikan (I4 boundary).
mokumo-shop is their correct home.

Binaries (mokumo-server, mokumo-desktop) repointed to
mokumo_shop::startup::* directly. services/api retains transitional
`pub use mokumo_shop::startup::*` re-exports so BDD tests continue to
compile until B3-bulk migrates them.

Refs #512 (Stage 8, part B2a).
