+++ crates/mokumo-shop/src/lib.rs
@@ -34,6 +34,7 @@ pub mod server_info;
 pub mod settings;
 pub mod setup;
 pub mod shop;
+pub mod startup;
 pub mod state;
 pub mod types;
 pub mod ws;
