+++ services/api/src/lib.rs
@@ -55,17 +55,8 @@ use kikan_types::HealthResponse;
 /// still reference the historic `mokumo_api::DEMO_RESET_PATH` path.
 pub use kikan::platform::auth::DEMO_RESET_PATH;
 
-/// Error returned by `setup_profile_db` and `prepare_database`.
-///
-/// Carries the human-readable error message and the path to the pre-migration backup
-/// (if one was created before the failure). The backup path lets callers surface the
-/// restore location to the shop owner in error dialogs and startup events.
-#[derive(Debug, thiserror::Error)]
-#[error("{message}")]
-pub struct ProfileDbError {
-    pub message: String,
-    pub backup_path: Option<std::path::PathBuf>,
-}
+/// Re-export from `mokumo_shop::startup` — lifted in PR 4b (#512).
+pub use mokumo_shop::startup::ProfileDbError;
 
 pub use kikan::platform::auth::PendingReset;
 
@@ -99,518 +90,11 @@ pub use mokumo_shop::profile_db_init::MokumoProfileDbInitializer;
 #[cfg(feature = "spa")]
 pub struct SpaAssets;
 
-/// Create the required data directories: data_dir, demo/, production/, and logs/.
-///
-/// Returns an error with the path included in the message on failure.
-pub fn ensure_data_dirs(data_dir: &Path) -> Result<(), std::io::Error> {
-    for dir in [
-        data_dir.to_path_buf(),
-        data_dir.join(SetupMode::Demo.as_dir_name()),
-        data_dir.join(SetupMode::Production.as_dir_name()),
-        data_dir.join("logs"),
-    ] {
-        std::fs::create_dir_all(&dir).map_err(|e| {
-            std::io::Error::new(
-                e.kind(),
-                format!("Failed to create directory {}: {}", dir.display(), e),
-            )
-        })?;
-    }
-    Ok(())
-}
-
-/// Read the `active_profile` file from the data directory.
-///
-/// Returns `SetupMode::Demo` if the file does not exist, is empty, or contains
-/// an unrecognised value (first launch defaults to demo).
-pub fn resolve_active_profile(data_dir: &Path) -> kikan::SetupMode {
-    use kikan::SetupMode;
-
-    let profile_path = data_dir.join("active_profile");
-    match std::fs::read_to_string(&profile_path) {
-        Ok(contents) => contents.trim().parse().unwrap_or(SetupMode::Demo),
-        Err(e) if e.kind() == std::io::ErrorKind::NotFound => SetupMode::Demo,
-        Err(e) => {
-            tracing::error!(path = %profile_path.display(), "Failed to read active_profile file: {e}; defaulting to demo");
-            SetupMode::Demo
-        }
-    }
-}
-
-/// Migrate a flat data directory layout to the dual-profile structure.
-///
-/// Idempotent: safe to call on every startup.
-///
-/// Steps:
-/// 1. If `production/mokumo.db` does NOT exist AND flat `mokumo.db` DOES exist:
-///    copy flat -> production/mokumo.db
-/// 2. If `active_profile` does NOT exist: write "production"
-///    (existing users who had a flat layout are production users)
-/// 3. If BOTH `production/mokumo.db` AND flat `mokumo.db` exist: remove flat
-pub fn migrate_flat_layout(data_dir: &Path) -> Result<(), std::io::Error> {
-    let flat_db = data_dir.join("mokumo.db");
-    let production_db = data_dir
-        .join(SetupMode::Production.as_dir_name())
-        .join("mokumo.db");
-    let profile_path = data_dir.join("active_profile");
-
-    let flat_exists = flat_db.try_exists()?;
-    let production_exists = production_db.try_exists()?;
-
-    // Step 1: Copy flat DB to production/ if production doesn't have one yet
-    if !production_exists && flat_exists {
-        std::fs::create_dir_all(data_dir.join(SetupMode::Production.as_dir_name()))?;
-        // Best-effort WAL checkpoint before copying: ensures committed but
-        // un-checkpointed transactions are included in the destination file.
-        // Logs a warning and continues if the file isn't in WAL mode or isn't
-        // a valid SQLite database (e.g., legacy installs that never used WAL).
-        if let Ok(conn) = rusqlite::Connection::open(&flat_db)
-            && let Err(e) = conn.execute_batch("PRAGMA wal_checkpoint(TRUNCATE)")
-        {
-            tracing::warn!(
-                "WAL checkpoint failed during flat DB migration (proceeding with copy): {e}"
-            );
-        }
-        std::fs::copy(&flat_db, &production_db)?;
-        tracing::info!("Migrated flat database to {}", production_db.display());
-    }
-
-    // Step 2: Write active_profile = "production" for existing users
-    if !profile_path.try_exists()? && flat_exists {
-        std::fs::write(&profile_path, SetupMode::Production.as_str())?;
-        tracing::info!("Set active profile to 'production' (migrated from flat layout)");
-    }
-
-    // Step 3: Remove flat DB — at this point production either already existed
-    // (crash recovery) or was just created by Step 1. Either way, flat is redundant.
-    if flat_exists {
-        std::fs::remove_file(&flat_db)?;
-        tracing::info!("Removed flat database after migration");
-        // WAL and SHM files may not exist — silently ignore NotFound
-        let _ = std::fs::remove_file(data_dir.join("mokumo.db-wal"));
-        let _ = std::fs::remove_file(data_dir.join("mokumo.db-shm"));
-    }
-
-    Ok(())
-}
-
-/// Shared startup sequence: create directories, migrate layout, copy sidecar,
-/// resolve profile, run guard chain and initialize both databases.
-///
-/// Guard chain per profile (active first, then non-active):
-///   1. `check_application_id` — reject non-Mokumo SQLite files
-///   2. `pre_migration_backup` — WAL-safe backup before any migration runs
-///   3. `check_schema_compatibility` — reject databases from newer Mokumo versions
-///      (demo: silent recreate from sidecar; production: hard abort)
-///   4. `initialize_database` — pool + PRAGMAs + `Migrator::up()`
-///
-/// Used by both the CLI server (`main.rs`) and the desktop app (`lib.rs`).
-/// Returns `(demo_db, production_db, profile)` on success.
-pub async fn prepare_database(
-    data_dir: &Path,
-) -> Result<(DatabaseConnection, DatabaseConnection, kikan::SetupMode), ProfileDbError> {
-    use kikan::SetupMode;
-
-    ensure_data_dirs(data_dir).map_err(|e| ProfileDbError {
-        message: format!("Failed to create data directories: {e}"),
-        backup_path: None,
-    })?;
-    migrate_flat_layout(data_dir).map_err(|e| ProfileDbError {
-        message: format!("Failed to migrate flat layout: {e}"),
-        backup_path: None,
-    })?;
-
-    if let Err(e) = kikan::platform::demo::copy_sidecar_if_needed(data_dir) {
-        tracing::warn!(
-            "Failed to copy demo sidecar: {e}; \
-             demo will start with empty database (no pre-seeded data)"
-        );
-    }
-
-    let profile = resolve_active_profile(data_dir);
-    let other_profile = match profile {
-        SetupMode::Demo => SetupMode::Production,
-        SetupMode::Production => SetupMode::Demo,
-    };
-
-    let active_db_path = data_dir.join(profile.as_str()).join("mokumo.db");
-    let other_db_path = data_dir.join(other_profile.as_str()).join("mokumo.db");
-
-    let active_db =
-        setup_profile_db(&active_db_path, profile == SetupMode::Production, data_dir).await?;
-    tracing::info!(
-        "Active database ({profile}) ready at {}",
-        active_db_path.display()
-    );
-
-    let other_db = setup_profile_db(
-        &other_db_path,
-        other_profile == SetupMode::Production,
-        data_dir,
-    )
-    .await?;
-    tracing::info!(
-        "Non-active database ({other_profile}) ready at {}",
-        other_db_path.display()
-    );
-
-    let (demo_db, production_db) = match profile {
-        SetupMode::Demo => (active_db, other_db),
-        SetupMode::Production => (other_db, active_db),
-    };
-
-    Ok((demo_db, production_db, profile))
-}
-
-/// Run the full guard chain for one profile database and return an initialized connection.
-///
-/// Guards run in order:
-///   1. `check_application_id` (pre-pool; only if DB file exists)
-///   2. `pre_migration_backup` (only if DB file exists)
-///   3. `ensure_auto_vacuum` (pre-pool; creates file for new DBs, VACUUMs existing ones)
-///   4. `check_schema_compatibility` (pre-pool; only if DB file exists)
-///      - If demo profile is incompatible: silently recreate from sidecar and continue.
-///      - If production profile is incompatible: hard abort with actionable message.
-///   5. `initialize_database` (pool + migrations)
-///
-/// Run `ensure_auto_vacuum` on a blocking thread and convert errors to `ProfileDbError`.
-async fn run_auto_vacuum_guard(
-    db_path: &Path,
-    backup_path: Option<std::path::PathBuf>,
-) -> Result<(), ProfileDbError> {
-    let db_path_owned = db_path.to_path_buf();
-    let display = db_path.display().to_string();
-    tokio::task::spawn_blocking(move || kikan::db::ensure_auto_vacuum(&db_path_owned))
-        .await
-        .map_err(|e| ProfileDbError {
-            message: format!("auto_vacuum guard panicked for {display}: {e}"),
-            backup_path: backup_path.clone(),
-        })?
-        .map_err(|e| ProfileDbError {
-            message: format!(
-                "Failed to enable auto_vacuum on {display}: {e}. \
-                 Check disk space (VACUUM requires ~2x database size).",
-            ),
-            backup_path,
-        })
-}
-
-/// Returns a human-readable error string on failure (technical detail sent to tracing).
-async fn setup_profile_db(
-    db_path: &Path,
-    is_production: bool,
-    data_dir: &Path,
-) -> Result<DatabaseConnection, ProfileDbError> {
-    use kikan::db::DatabaseSetupError;
-
-    // Pre-migration backup only runs when the DB file already exists.
-    // Track this so format_db_setup_error can omit the backup claim for fresh installs.
-    let backup_taken = db_path.exists();
-    // Capture the backup path from Guard 2 so all subsequent guard failures can include it.
-    let mut backup_path: Option<std::path::PathBuf> = None;
-
-    if backup_taken {
-        // Guard 1: confirm this file belongs to Mokumo
-        kikan::db::check_application_id(db_path).map_err(|e| match e {
-            DatabaseSetupError::NotKikanDatabase { ref path } => ProfileDbError {
-                message: format!(
-                    "The database at {} is not a Mokumo database. \
-                     Check your --data-dir setting.",
-                    path.display()
-                ),
-                backup_path: None, // Guard 1 fires before Guard 2; no backup yet.
-            },
-            _ => ProfileDbError {
-                message: format!("application_id check failed for {}: {e}", db_path.display()),
-                backup_path: None,
-            },
-        })?;
-
-        // Guard 2: backup before any migration runs
-        backup_path = kikan::backup::pre_migration_backup(db_path)
-            .await
-            .map_err(|e| ProfileDbError {
-                message: format!(
-                    "Pre-migration backup failed for {}: {e}. \
-                     Refusing to run migrations without a backup. \
-                     Check disk space and permissions.",
-                    db_path.display()
-                ),
-                backup_path: None,
-            })?;
-
-        // Guard 2b: ensure auto_vacuum = INCREMENTAL (one-time VACUUM if needed)
-        run_auto_vacuum_guard(db_path, backup_path.clone()).await?;
-
-        // Guard 3: reject databases from newer Mokumo versions
-        match mokumo_shop::db::check_schema_compatibility(db_path) {
-            Ok(()) => {}
-            Err(DatabaseSetupError::SchemaIncompatible {
-                ref path,
-                ref unknown_migrations,
-            }) => {
-                if is_production {
-                    return Err(ProfileDbError {
-                        message: format!(
-                            "The production database at {} was created by a newer version of Mokumo. \
-                             Please upgrade Mokumo to the latest version, or restore from a backup. \
-                             Do not delete the database — your production data is there.",
-                            path.display()
-                        ),
-                        backup_path: backup_path.clone(),
-                    });
-                }
-                // Demo profile: silent recreate from sidecar (ephemeral data)
-                tracing::warn!(
-                    ?unknown_migrations,
-                    "Demo database has unknown migrations from newer Mokumo version; \
-                     resetting to fresh demo data."
-                );
-                kikan::platform::demo::force_copy_sidecar(data_dir).map_err(|e| {
-                    ProfileDbError {
-                        message: format!("Failed to reset demo database: {e}"),
-                        backup_path: backup_path.clone(),
-                    }
-                })?;
-                // Re-run guards on the fresh sidecar before initializing.
-                // The bundled sidecar could theoretically be malformed or from a future version.
-                // The sidecar's own pre_migration_backup result is discarded — the original
-                // backup_path (from the user's old demo data) is not relevant here.
-                if db_path.exists() {
-                    kikan::db::check_application_id(db_path).map_err(|e| match e {
-                        DatabaseSetupError::NotKikanDatabase { ref path } => ProfileDbError {
-                            message: format!(
-                                "The bundled demo database is not a valid Mokumo database: {}. \
-                                 Please reinstall Mokumo.",
-                                path.display()
-                            ),
-                            backup_path: None,
-                        },
-                        _ => ProfileDbError {
-                            message: format!(
-                                "application_id check failed for demo database after reset: {e}"
-                            ),
-                            backup_path: None,
-                        },
-                    })?;
-                    let _sidecar_backup = kikan::backup::pre_migration_backup(db_path)
-                        .await
-                        .map_err(|e| ProfileDbError {
-                            message: format!(
-                                "Pre-migration backup failed for demo database after reset: {e}"
-                            ),
-                            backup_path: None,
-                        })?;
-                    // Guard 2b on sidecar: ensure auto_vacuum
-                    run_auto_vacuum_guard(db_path, None).await?;
-                    if let Err(e) = mokumo_shop::db::check_schema_compatibility(db_path) {
-                        return Err(ProfileDbError {
-                            message: format!(
-                                "Demo database failed compatibility check after reset: {e}"
-                            ),
-                            backup_path: None,
-                        });
-                    }
-                }
-                let url = format!("sqlite:{}?mode=rwc", db_path.display());
-                return mokumo_shop::db::initialize_database(&url)
-                    .await
-                    .map_err(|e| ProfileDbError {
-                        message: format_db_setup_error(e, db_path, true),
-                        backup_path: None,
-                    });
-            }
-            Err(e) => {
-                return Err(ProfileDbError {
-                    message: format!(
-                        "Schema compatibility check failed for {}: {e}",
-                        db_path.display()
-                    ),
-                    backup_path: backup_path.clone(),
-                });
-            }
-        }
-    } else {
-        // New database: ensure auto_vacuum is set before pool creation
-        run_auto_vacuum_guard(db_path, None).await?;
-    }
-
-    // Guard 4: initialize pool + run migrations
-    let url = format!("sqlite:{}?mode=rwc", db_path.display());
-    mokumo_shop::db::initialize_database(&url)
-        .await
-        .map_err(|e| ProfileDbError {
-            message: format_db_setup_error(e, db_path, backup_taken),
-            backup_path: backup_path.clone(),
-        })
-}
-
-/// Format a `DatabaseSetupError` into a human-readable message for the operator.
-///
-/// `backup_taken` indicates whether `pre_migration_backup` ran before the error occurred.
-/// When `false` (fresh install), the backup claim is omitted to avoid a false assertion.
-///
-/// Technical details (DbErr internals) are sent to `tracing::error!` only.
-fn format_db_setup_error(
-    e: kikan::db::DatabaseSetupError,
-    db_path: &Path,
-    backup_taken: bool,
-) -> String {
-    use kikan::db::DatabaseSetupError;
-    tracing::error!("Database setup error for {}: {:?}", db_path.display(), e);
-    match e {
-        DatabaseSetupError::Migration(_) => {
-            let backup_note = if backup_taken {
-                " Your data was backed up before the attempt."
-            } else {
-                ""
-            };
-            format!(
-                "Mokumo could not apply a database migration to {}.{backup_note} \
-                 Contact support if this persists.",
-                db_path.display()
-            )
-        }
-        DatabaseSetupError::SchemaIncompatible { ref path, .. } => format!(
-            "The database at {} was created by a newer version of Mokumo. \
-             Please upgrade Mokumo to the latest version, or restore from a backup.",
-            path.display()
-        ),
-        DatabaseSetupError::NotKikanDatabase { ref path } => format!(
-            "The database at {} is not a Mokumo database. \
-             Check your --data-dir setting.",
-            path.display()
-        ),
-        DatabaseSetupError::Pool(_) => format!(
-            "Failed to open database connection pool at {}. \
-             Check disk space and file permissions.",
-            db_path.display()
-        ),
-        DatabaseSetupError::Query(_) => format!(
-            "A database query failed during initialization of {}. \
-             Your data was not modified.",
-            db_path.display()
-        ),
-        DatabaseSetupError::Rusqlite(_) => format!(
-            "A low-level database error occurred while checking {}. \
-             Check disk space and file permissions.",
-            db_path.display()
-        ),
-    }
-}
-
-/// Attempt to bind a TCP listener, trying ports from `port` through `port + 10`.
-///
-/// Returns the listener and the actual port that was bound. Logs at INFO when
-/// a port is successfully bound. Returns an error if all 11 ports are exhausted.
-pub async fn try_bind(
-    host: &str,
-    port: u16,
-) -> Result<(tokio::net::TcpListener, u16), std::io::Error> {
-    let end_port = port.saturating_add(10);
-    for p in port..=end_port {
-        let addr = format!("{host}:{p}");
-        match tokio::net::TcpListener::bind(&addr).await {
-            Ok(listener) => {
-                let actual_port = listener.local_addr()?.port();
-                tracing::info!("Listening on {host}:{actual_port}");
-                return Ok((listener, actual_port));
-            }
-            Err(e) if e.kind() == std::io::ErrorKind::AddrInUse => {
-                tracing::debug!("Port {p} in use, trying next");
-            }
-            Err(e) => {
-                return Err(std::io::Error::new(
-                    e.kind(),
-                    format!("Cannot bind to {host}:{p}: {e}"),
-                ));
-            }
-        }
-    }
-    Err(std::io::Error::new(
-        std::io::ErrorKind::AddrInUse,
-        format!(
-            "All ports {port}-{end_port} are occupied. \
-             Use --port to specify a different port, or close conflicting applications."
-        ),
-    ))
-}
-
-// `try_bind_ephemeral_loopback` was lifted to `kikan-tauri::net` in S4.2 (#507).
-// The ephemeral-loopback bind strategy is specific to the Tauri shell and belongs
-// in the Tauri adapter crate (I2 hygiene). services/api deliberately does NOT
-// re-export from kikan-tauri — doing so would drag `tauri` into `mokumo-server`
-// via the mokumo-api dep, violating I3 (headless binary has no Tauri deps).
-// Desktop callers import `kikan_tauri::try_bind_ephemeral_loopback` directly.
-
-/// Generate a random setup token (UUID v4).
-pub fn generate_setup_token() -> String {
-    uuid::Uuid::new_v4().to_string()
-}
-
-/// Shared init for session store + setup state. Used by both `build_app` and
-/// `build_app_with_shutdown` to avoid duplicating the setup/session bootstrap.
-///
-/// The session store is opened from a SEPARATE SQLite database at `session_db_path`,
-/// keeping session data independent of the active profile database.
-///
-/// `production_db` is used to check setup completion status (production setup
-/// is what the wizard populates).
-pub async fn init_session_and_setup(
-    production_db: &DatabaseConnection,
-    session_db_path: &Path,
-) -> Result<(SqliteStore, Arc<AtomicBool>, Option<String>), Box<dyn std::error::Error + Send + Sync>>
-{
-    let is_complete = mokumo_shop::db::is_setup_complete(production_db).await?;
-    let setup_completed = Arc::new(AtomicBool::new(is_complete));
-    let setup_token = if is_complete {
-        None
-    } else {
-        Some(generate_setup_token())
-    };
-
-    // Open a separate SQLite pool for sessions
-    let session_url = format!("sqlite:{}?mode=rwc", session_db_path.display());
-    let session_pool = kikan::db::open_raw_sqlite_pool(&session_url)
-        .await
-        .map_err(|e| {
-            format!(
-                "Failed to open session database at {}: {e}",
-                session_db_path.display()
-            )
-        })?;
-    let session_store = SqliteStore::new(session_pool);
-    session_store
-        .migrate()
-        .await
-        .map_err(|e| format!("Session store migration failed: {e}"))?;
-
-    Ok((session_store, setup_completed, setup_token))
-}
-
-/// Build the Axum router with health check, SPA fallback, and tracing.
-///
-/// Resolve the `demo_install_ok` flag at startup.
-///
-/// Runs `validate_installation` against the demo DB when the active profile is Demo;
-/// always returns `true` for Production (an empty production DB is valid — setup is
-/// pending, not broken). Logs the result at `info` level for observability.
-pub async fn resolve_demo_install_ok(
-    demo_db: &DatabaseConnection,
-    active_profile: SetupMode,
-) -> Arc<AtomicBool> {
-    let ok = if active_profile == SetupMode::Demo {
-        let ok = kikan::db::validate_installation(demo_db).await;
-        tracing::info!(
-            demo_install_ok = ok,
-            "demo installation validation complete"
-        );
-        ok
-    } else {
-        true
-    };
-    Arc::new(AtomicBool::new(ok))
-}
+/// Re-exports from `mokumo_shop::startup` — lifted in PR 4b (#512).
+pub use mokumo_shop::startup::{
+    ensure_data_dirs, generate_setup_token, init_session_and_setup, migrate_flat_layout,
+    prepare_database, resolve_active_profile, resolve_demo_install_ok, try_bind,
+};
 
 /// Test-only convenience wrapper. Does NOT spawn the background IP refresh
 /// task — the local IP is computed once and never updated. Use
